move model into its own file
